- make callWithBaggage static - extract Propagation class with static methods
